Switch to Adafruit_GPIO for cross-platform compatibility
This should allow the library to work on the Raspberry Pi, Beaglebone Black, and Intel Edison/Galileo boards now.

It's only been tested on the RPi3 at the moment. The main motivation
for switching was actually that the Adafruit library provides an
abstraction around the `RPi.GPIO` library, which makes testing/mocking
easier instead of mucking with mocking imports.